requirements.txt
pandas==1.5.2
black==22.12.0
fastapi==0.88.0
kaggle==1.5.12
uvicorn==0.20.0
psycopg2-binary==2.9.5